LayoutTests/websocket/tests/handshake-fail-by-sub-protocol-mismatch-expected.txt

CONSOLE MESSAGE: line 0: Error during WebSocket handshake: protocol mismatch: requestting-protocol != available-protocol
Test WebSocket handshake fail if sub protocol name mismatches.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS protocol is undefined.
PASS successfullyParsed is true

TEST COMPLETE
